Former-commit-id: f7f97d9f5442e7c26bedf7fde3ad9c57738f0087 [formerly 346d775f57d1a426208857a6bb2da5f06d3e48c0 [formerly 25026a4d9955b76039bf25861380cadebff19665]]
Former-commit-id: 346d775f57d1a426208857a6bb2da5f06d3e48c0
Former-commit-id: e372481f81d2b16993a35e6fb0a10d937b617ffc

+++ edexOsgi/com.raytheon.uf.common.geospatial/src/com/raytheon/uf/common/geospatial/interpolation/data/AbstractDataWrapper.java
@@ -86,7 +86,7 @@ public abstract class AbstractDataWrapper implements DataSource,
     public double getDataValue(int x, int y) {
         if (y < 0 || y > ny - 1) {
             // outside y range
-            return Float.NaN;
+            return Double.NaN;
         } else if (x < 0 || x > nx - 1) {
             // outside x range
             if (wrapX > 0) {
